Change `as_s32()` and `as_u32()` to be silent
Diffs from this change in OOT, MM, PM:

https://gist.github.com/zbanks/d8655acae883aa80a01c265c7acde0f8

+++ tests/end_to_end/gcc-division/gcc-o2-out.c
@@ -15,14 +15,14 @@ void test(u32 a) {
 }
 
 void func_00400098(s8 arg0) {
-    func_00400090((u32) ((s32) ((s8) subroutine_arg4 + ((u32) (subroutine_arg4 << 0x18) >> 0x1F)) >> 1), arg0);
+    func_00400090((u32) (((s8) subroutine_arg4 + ((subroutine_arg4 << 0x18) >> 0x1F)) >> 1), arg0);
     func_00400090((u32) (s8) (subroutine_arg4 / 3));
     func_00400090((u32) (s8) (subroutine_arg4 / 5));
     func_00400090((u32) (s8) (subroutine_arg4 / 7));
     func_00400090((u32) (s8) (subroutine_arg4 / 10));
     func_00400090((u32) (s8) (subroutine_arg4 / 100));
     func_00400090(subroutine_arg4 / 255);
-    func_00400090((u32) (s8) ((s8) subroutine_arg4 - (((s32) ((s8) subroutine_arg4 + ((u32) (subroutine_arg4 << 0x18) >> 0x1F)) >> 1) * 2)));
+    func_00400090((u32) (s8) ((s8) subroutine_arg4 - ((((s8) subroutine_arg4 + ((subroutine_arg4 << 0x18) >> 0x1F)) >> 1) * 2)));
     func_00400090((u32) (s8) (subroutine_arg4 % 3));
     func_00400090((u32) (s8) (subroutine_arg4 % 5));
     func_00400090((u32) (s8) (subroutine_arg4 % 7));
@@ -32,7 +32,7 @@ void func_00400098(s8 arg0) {
 }
 
 void func_004003A8(s16 arg0) {
-    func_00400090((u32) ((s32) ((s16) subroutine_arg4 + ((u32) (subroutine_arg4 << 0x10) >> 0x1F)) >> 1), arg0);
+    func_00400090((u32) (((s16) subroutine_arg4 + ((subroutine_arg4 << 0x10) >> 0x1F)) >> 1), arg0);
     func_00400090((u32) (s16) (subroutine_arg4 / 3));
     func_00400090((u32) (s16) (subroutine_arg4 / 5));
     func_00400090((u32) (s16) (subroutine_arg4 / 7));
@@ -41,7 +41,7 @@ void func_004003A8(s16 arg0) {
     func_00400090((u32) (s16) (subroutine_arg4 / 255));
     func_00400090((u32) (s16) (subroutine_arg4 / 360));
     func_00400090(subroutine_arg4 / 65534);
-    func_00400090((u32) (s16) ((s16) subroutine_arg4 - (((s32) ((s16) subroutine_arg4 + ((u32) (subroutine_arg4 << 0x10) >> 0x1F)) >> 1) * 2)));
+    func_00400090((u32) (s16) ((s16) subroutine_arg4 - ((((s16) subroutine_arg4 + ((subroutine_arg4 << 0x10) >> 0x1F)) >> 1) * 2)));
     func_00400090((u32) (s16) (subroutine_arg4 % 3));
     func_00400090((u32) (s16) (subroutine_arg4 % 5));
     func_00400090((u32) (s16) (subroutine_arg4 % 7));
@@ -61,7 +61,7 @@ void func_004007C0(u32 arg0) {
     s32 phi_a0_6;
 
     func_00400090(arg0);
-    func_00400090((u32) ((s32) (arg0 + (arg0 >> 0x1F)) >> 1));
+    func_00400090((u32) ((arg0 + (arg0 >> 0x1F)) >> 1));
     func_00400090(arg0 / 3);
     phi_a0 = (s32) arg0;
     if ((s32) arg0 < 0) {
@@ -139,9 +139,9 @@ void func_004007C0(u32 arg0) {
     if ((s32) arg0 < 0) {
         phi_a0_6 = arg0 + 3;
     }
-    func_00400090((u32) -(s32) (phi_a0_6 >> 2));
+    func_00400090((u32) -(phi_a0_6 >> 2));
     func_00400090(arg0 / -3);
-    func_00400090((u32) -(s32) ((s32) (arg0 + (arg0 >> 0x1F)) >> 1));
+    func_00400090((u32) -((arg0 + (arg0 >> 0x1F)) >> 1));
     func_00400090((u32) -(s32) arg0);
 }
 
@@ -253,7 +253,7 @@ void func_00401AA0(u32 arg0) {
     func_00400090(arg0 / 5);
     func_00400090(arg0 / 6);
     temp_hi = arg0 / 7;
-    func_00400090((u32) (temp_hi + ((u32) (arg0 - temp_hi) >> 1)) >> 2);
+    func_00400090((temp_hi + ((arg0 - temp_hi) >> 1)) >> 2);
     func_00400090(arg0 >> 3);
     func_00400090(arg0 / 9);
     func_00400090(arg0 / 10);
@@ -266,22 +266,22 @@ void func_00401AA0(u32 arg0) {
     func_00400090(arg0 / 17);
     func_00400090(arg0 / 18);
     temp_hi_2 = MULTU_HI(arg0, 0xAF286BCB);
-    func_00400090((u32) (temp_hi_2 + ((u32) (arg0 - temp_hi_2) >> 1)) >> 4);
+    func_00400090((temp_hi_2 + ((arg0 - temp_hi_2) >> 1)) >> 4);
     func_00400090(arg0 / 20);
     temp_hi_3 = MULTU_HI(arg0, 0x86186187);
-    func_00400090((u32) (temp_hi_3 + ((u32) (arg0 - temp_hi_3) >> 1)) >> 4);
+    func_00400090((temp_hi_3 + ((arg0 - temp_hi_3) >> 1)) >> 4);
     func_00400090(arg0 / 22);
     func_00400090(arg0 / 23);
     func_00400090(arg0 / 24);
     func_00400090(arg0 / 25);
     func_00400090(arg0 / 26);
     temp_hi_4 = MULTU_HI(arg0, 0x2F684BDB);
-    func_00400090((u32) (temp_hi_4 + ((u32) (arg0 - temp_hi_4) >> 1)) >> 4);
+    func_00400090((temp_hi_4 + ((arg0 - temp_hi_4) >> 1)) >> 4);
     func_00400090(arg0 / 28);
     func_00400090(arg0 / 29);
     func_00400090(arg0 / 30);
     temp_hi_5 = arg0 / 31;
-    func_00400090((u32) (temp_hi_5 + ((u32) (arg0 - temp_hi_5) >> 1)) >> 4);
+    func_00400090((temp_hi_5 + ((arg0 - temp_hi_5) >> 1)) >> 4);
     func_00400090(arg0 >> 5);
     func_00400090(arg0 / 33);
     func_00400090(arg0 / 100);
@@ -294,10 +294,10 @@ void func_00401AA0(u32 arg0) {
     func_00400090(arg0 / 10000000);
     func_00400090(arg0 / 100000000);
     func_00400090(arg0 >> 0x1E);
-    func_00400090((u32) MULTU_HI(arg0, -3) >> 0x1E);
+    func_00400090(MULTU_HI(arg0, -3) >> 0x1E);
     func_00400090(arg0 / 2147483645);
     temp_hi_6 = arg0 / 1431655765;
-    func_00400090((u32) (temp_hi_6 + ((u32) (arg0 - temp_hi_6) >> 1)) >> 0x1E);
+    func_00400090((temp_hi_6 + ((arg0 - temp_hi_6) >> 1)) >> 0x1E);
     func_00400090(arg0 >> 0x1F);
     func_00400090((arg0 < 0x80000001U) ^ 1);
     func_00400090((arg0 < 0xFFFFFFFEU) ^ 1);
@@ -319,7 +319,7 @@ void func_00401FC4(u32 arg0) {
     func_00400090(arg0 / 5);
     func_00400090(arg0 / 6);
     temp_hi = arg0 / 7;
-    func_00400090((u32) (temp_hi + ((u32) (arg0 - temp_hi) >> 1)) >> 2);
+    func_00400090((temp_hi + ((arg0 - temp_hi) >> 1)) >> 2);
     func_00400090(arg0 >> 3);
     func_00400090(arg0 / 9);
     func_00400090(arg0 / 10);
@@ -332,22 +332,22 @@ void func_00401FC4(u32 arg0) {
     func_00400090(arg0 / 17);
     func_00400090(arg0 / 18);
     temp_hi_2 = MULTU_HI(arg0, 0xAF286BCB);
-    func_00400090((u32) (temp_hi_2 + ((u32) (arg0 - temp_hi_2) >> 1)) >> 4);
+    func_00400090((temp_hi_2 + ((arg0 - temp_hi_2) >> 1)) >> 4);
     func_00400090(arg0 / 20);
     temp_hi_3 = MULTU_HI(arg0, 0x86186187);
-    func_00400090((u32) (temp_hi_3 + ((u32) (arg0 - temp_hi_3) >> 1)) >> 4);
+    func_00400090((temp_hi_3 + ((arg0 - temp_hi_3) >> 1)) >> 4);
     func_00400090(arg0 / 22);
     func_00400090(arg0 / 23);
     func_00400090(arg0 / 24);
     func_00400090(arg0 / 25);
     func_00400090(arg0 / 26);
     temp_hi_4 = MULTU_HI(arg0, 0x2F684BDB);
-    func_00400090((u32) (temp_hi_4 + ((u32) (arg0 - temp_hi_4) >> 1)) >> 4);
+    func_00400090((temp_hi_4 + ((arg0 - temp_hi_4) >> 1)) >> 4);
     func_00400090(arg0 / 28);
     func_00400090(arg0 / 29);
     func_00400090(arg0 / 30);
     temp_hi_5 = arg0 / 31;
-    func_00400090((u32) (temp_hi_5 + ((u32) (arg0 - temp_hi_5) >> 1)) >> 4);
+    func_00400090((temp_hi_5 + ((arg0 - temp_hi_5) >> 1)) >> 4);
     func_00400090(arg0 >> 5);
     func_00400090(arg0 / 33);
     func_00400090(arg0 / 100);
@@ -360,10 +360,10 @@ void func_00401FC4(u32 arg0) {
     func_00400090(arg0 / 10000000);
     func_00400090(arg0 / 100000000);
     func_00400090(arg0 >> 0x1E);
-    func_00400090((u32) MULTU_HI(arg0, -3) >> 0x1E);
+    func_00400090(MULTU_HI(arg0, -3) >> 0x1E);
     func_00400090(arg0 / 2147483645);
     temp_hi_6 = arg0 / 1431655765;
-    func_00400090((u32) (temp_hi_6 + ((u32) (arg0 - temp_hi_6) >> 1)) >> 0x1E);
+    func_00400090((temp_hi_6 + ((arg0 - temp_hi_6) >> 1)) >> 0x1E);
     func_00400090(arg0 >> 0x1F);
     func_00400090((arg0 < 0x80000001U) ^ 1);
     func_00400090((arg0 < 0xFFFFFFFEU) ^ 1);
